ZJIT: Move EP index/offset conversion functions to cruby
Keep the functions together (not split in HIR/codegen) and in a more
reasonable module.
